Extract shared code in prep for more computed metrics
* Extracts test factory methods for use in other tests.
* Extracts some helper queries for grabbing relevant token accounts
