+++ src/computed_metrics/wallet_metrics.ts
@@ -1,5 +1,6 @@
 import { getKnex } from "..//database";
 import { TokenAccountMint } from "../knex-types/token_account_mint";
+import { tokenDatabaseIds } from "./utils";
 
 const knex = getKnex();
 
@@ -9,14 +10,10 @@ export async function totalWallets(
 ) {
   const startDateFilter = opts?.startDate || new Date(0);
 
-  const relevantTokenIds = mintedTokens
-    .map((t) => t.id)
-    .filter((id) => id != null) as number[];
-
   const validAccounts = knex
     .select("token_account_id")
     .from("token_accounts")
-    .whereIn("mint_id", relevantTokenIds);
+    .whereIn("mint_id", tokenDatabaseIds(mintedTokens));
 
   const result = await knex("token_account_balances")
     .countDistinct("token_account_id")
